+++ CHANGELOG.md
@@ -5,6 +5,8 @@
 ### Changed
 
 - Upgrade `broadway/broadway` version
+- Changed factory method to no longer require the database name, instead you pass it the name of the collection you want to project
+into and an optional custom repository class.
 
 ## 0.1.0 - 2016-04-13